# PERF/REGR: astype changing order of some 2d data