--- a/pandas/tests/frame/methods/test_astype.py
+++ b/pandas/tests/frame/methods/test_astype.py
@@ -670,10 +670,10 @@ class TestAstype:
         result = DataFrame(["foo", "bar", "baz"]).astype(bytes)
         assert result.dtypes[0] == np.dtype("S3")
 
-    @pytest.mark.parametrize("step1,step2", [(6, 3), (6, 1), (1, 3)])
+    @pytest.mark.parametrize("step1,step2", [(2, 2), (2, 1), (1, 2)])
     def test_astype_noncontiguous(self, step1, step2):
         # GH#42396
-        data = np.arange(72).reshape(12, 6)
+        data = np.arange(16).reshape(4, 4)
         df = DataFrame(data)
 
         result = df.iloc[:step1, :step2].astype("int32").astype("int64")