@mzeitlin11 2021-07-13
> im a bit confused as to what parts of this are about a regression[bug] vs a regression[perf]. @mzeitlin11 will this be obvious once i get some caffeine?

Haha, I'll try to clarify :)

- Bug: The `ravel` and `astype` could happen with different orders, garbling the data. This can be fixed by specifying the same order in `ravel` as used for `astype`. However, this will not fix the perf regression
- Perf: With behavior fix above, we can end up with F-contiguous data depending on what contiguity goes into the `astype` call. For the reduction regressions at the top of the issue, this F-contiguity is what caused the regression - the routines are just slower on F-contig data. Just leaving default order ensures C-contiguity, fixing the perf regression.