@jreback 2021-07-13
@meeseeksdev backport 1.3.x